requirements.txt
flask
flask-cors
Flask-Migrate
Flask-SQLAlchemy
python-dotenv
psycopg2-binary
sqlalchemy
redis
celery
